skillTest/back/abm/requirements.txt
boto3==1.10.6
botocore==1.13.6
certifi==2019.9.11
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.6
django-braces==1.13.0
django-crispy-forms==1.8.0
django-extensions==2.2.5
django-heroku==0.3.1
djangorestframework==3.11.0
docutils==0.15.2
ecdsa==0.13.3
envs==1.3
future==0.18.2
gunicorn==20.0.4
idna==2.8
jmespath==0.9.4
python-dateutil==2.8.0
psycopg2==2.7.5
pytz==2019.3
requests==2.22.0
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.6
whitenoise==5.0.1